crud 에러 수지ㅓㅇ
