+++ api/calendarAPI.js
@@ -16,37 +16,4 @@ const calendarAPI = new Calendar('#calendar', {
         },
     },
 });
-
-//캐린더 이벤트 핸들링
-const event = calendarAPI.on({
-    clickSchedule: function (e) {
-        console.log('clickSchedule', e);
-    },
-    beforeCreateSchedule: function (e) {
-        console.log('beforeCreateSchedule', e);
-        const newDataObj = {
-            calendarId: '1',
-            isReadOnly: true,
-            isAllDay: e.isAllDay,
-            end: new Date(e.end.getTime()).toJSON(),
-            start: new Date(e.end.getTime()).toJSON(),
-            title: e.title,
-            location: e.location,
-        };
-        calendarAPI.createSchedules([newDataObj]);
-        console.log(calendarAPI.getSchedule());
-    },
-    beforeUpdateSchedule: function (e) {
-        console.log('beforeUpdateSchedule', e);
-        e.schedule.start = e.start;
-        e.schedule.end = e.end;
-        cal.updateSchedule(e.schedule.id, e.schedule.calendarId, e.schedule);
-    },
-    beforeDeleteSchedule: function (e) {
-        console.log('beforeDeleteSchedule', e);
-        cal.deleteSchedule(e.schedule.id, e.schedule.calendarId);
-    },
-});
-
-export { event };
 export default calendarAPI;
